Fix window icon on Linux (#188)
The `icon` parameter of windows (only used when the platform is Linux)
was supplied an image with dimensions of over 2000x2000. As a result of
the large dimensions, the icon would not be used. This commit resizes
the icon to 512x512, a size that allows the icon to be used in the
operating system and in the application itself.